feat: 2차 모듈구현(아이콘 활성화/비활성화 파트)
